+++ src/Exceptions/InvalidCollectionException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Anteris\DataTransferObjectFactory\Exceptions;
+
+class InvalidCollectionException extends \Exception
+{
+}