+++ src/Exceptions/InvalidDataTransferObjectException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Anteris\DataTransferObjectFactory\Exceptions;
+
+class InvalidDataTransferObjectException extends \Exception
+{
+}